improve error issue creation
