+++ packages/altair-app/src/app/modules/altair/error-handler.ts
@@ -4,6 +4,8 @@ import { debug } from './utils/logger';
 import { NotifyService } from './services';
 import newGithubIssueUrl from 'new-github-issue-url';
 import { UnknownError } from './interfaces/shared';
+import { truncateText } from './utils';
+import { getIssueUrl } from './utils/issue';
 
 @Injectable()
 export class GlobalErrorHandler implements ErrorHandler {
@@ -17,20 +19,8 @@ export class GlobalErrorHandler implements ErrorHandler {
     } else {
       const notifyService = this.injector.get(NotifyService);
       const errorMessage = error.message ? error.message : error.toString();
-      const issueTitle = `Bug: Application error: ${errorMessage}`;
-      const issueBody = `
-        \nError message: ${errorMessage}\n\nError stack:\n\n\`\`\`\n${error.stack}\n\`\`\`
-        \n### When did this error occur? And does it re occur, or it happened only once?\n
-        \n### Please provide some helpful screenshots of the bug
-      `;
-      const issueUrl = newGithubIssueUrl({
-        user: 'imolorhe',
-        repo: 'altair',
-        title: issueTitle,
-        body: issueBody,
-        labels: [ 'bug-report' ],
-        template: 'Bug_report.md',
-      });
+
+      const issueUrl = getIssueUrl(error);
       debug.error('Application error:', errorMessage);
       notifyService.error(`An error occured: ${errorMessage}`);
       notifyService.warning(`If you think this is a bug, click here to report the bug.`, 'Altair', {
